Add Predefined Formatters to TimeTag.format
Now, before assuming that formatText is a valid pattern, we check to see if formatText is a valid formatter instead. We get the field, then attempt to get the value from that field. If formatText isn't a valid formatter, then it will throw a noSuchField exception before it gets to the .get() (based on my very limited understanding of java lol). If we catch the exception, we instead set the format to formatText, which was basically the original behaviour.